Move builders into their own folders
